sleutelkastje/requirements.txt
Flask==2.3.2
Flask-HTTPAuth==4.8.0
Flask-pyoidc==3.10.0
psycopg2-binary==2.9.6
jinja2==3.1.2
gunicorn==20.1.0
flask-sqlalchemy
flask-migrate
